Removed EntityType.h , created namespace Renderer, improved Leaderboard, now it saves score after each game

+++ Source/Utility.cpp
@@ -18,16 +18,14 @@ bool CheckCollision(Vector2 circleOrigin, float circleRadius, Vector2 A, Vector2
 	{
 		return true;
 	}
-
 	constexpr float buffer = 0.1f;
 	const float length = lineLength(A, B);
-	const float dotP = static_cast<float>((((circleOrigin.x - A.x) * (B.x - A.x)) + ((circleOrigin.y - A.y) * (B.y - A.y))) / pow(length, 2));
+	const float dotP = (((circleOrigin.x - A.x) * (B.x - A.x)) + ((circleOrigin.y - A.y) * (B.y - A.y))) / pow(length, 2);
 	const float closestX = A.x + (dotP * (B.x - A.x));
 	const float closestY = A.y + (dotP * (B.y - A.y));
 	const float closeToStart = lineLength(A, { closestX, closestY });
 	const float closeToEnd = lineLength(B, { closestX, closestY });
 	const float closestLength = closeToStart + closeToEnd;
-
 	if (closestLength == length + buffer || closestLength == length - buffer)
 	{
 		return closeToStart < circleRadius;
